Base: On-The-Go mode squashed [1/2]
On-The-Go turns on the camera of your phone and streams it to an overlay,
which is drawn over everything.
this makes you feel like your display is transparent, as you see the content
behind it.

its really useful when you are using your phone while walking, thats why
we called it "On-The-Go"

allows to change transparency of the overlay via the notification.

have fun!

Change-Id: I3efccb5fac41a946f6cb4992e9181422980ee5e4

b0b3c1a Fix Derps
01ec6d7 On-The-Go: make it multiuser aware
c3f233c On-The-Go: handle display orientation changes
78cb7b7 Base: On-The-Go: Turn stop and start service if screen turns off and on
8d65360 Base: On-The-Go: Hide on devices without camera [1/2]
aeb8d59 Base: On-The-Go: Switch Licenses to GPLv3
8a820c5 Base: On-The-Go: Add quicktile [1/2]
5ce7ff2 Base: On-The-Go: Allow to restart service on the fly
1c19b75 Base: On-The-Go: Change the way of changing the camera mode
caf4510 Base: On-The-Go: Overlay status- and navigation bar as well
94aa412 Base: On-The-Go: Several fixes
055cfc7 Base: On-The-Go: Allow to toggle between front and back camera
b97c331 Base: On-The-Go mode [1/2]